:wrench: check if is not an arch linux

+++ install/install_for_macos_linux.sh
@@ -14,11 +14,13 @@ warning() {
     DISTRO_NAME=$(cat /etc/*elease | grep DISTRIB_ID | cut -d '=' -f2)
     DISTRO_VERSION=$(cat /etc/*elease | grep DISTRIB_RELEASE | cut -d '=' -f2 | sed 's@^[^0-9]*\([0-9]\+\).*@\1@')
     if [[ -z "$DISTRO_VERSION" ]] || [[ -z "$DISTRO_NAME" ]] || ( [ "$DISTRO_VERSION" -lt 20 ] && [ "$DISTRO_NAME" != "Ubuntu" ]);then
-      warning "tipi is currently supported on Ubuntu 20.04 or later only. You are running an unsuported distribution."
-      warning "Do you want to install tipi anyway? (y/n)"
-      read answer
-      if [ "$answer" == "${answer#[Yy]}" ] ;then 
-       exit 1
+      if [ ! -f "/etc/arch-release" ]; then
+        warning "tipi is currently supported on Ubuntu 20.04 or later only. You are running an unsuported distribution."
+        warning "Do you want to install tipi anyway? (y/n)"
+        read answer
+        if [ "$answer" == "${answer#[Yy]}" ] ;then 
+        exit 1
+        fi
       fi
     fi
   elif [ "$(uname)" == "Darwin" ]; then
